+++ CodeMaid/Integration/Commands/JoinLinesCommand.cs
@@ -1,14 +1,26 @@
 using EnvDTE;
 using SteveCadwallader.CodeMaid.Helpers;
-using System.ComponentModel.Design;
+using SteveCadwallader.CodeMaid.Properties;
 
 namespace SteveCadwallader.CodeMaid.Integration.Commands
 {
     /// <summary>
     /// A command that provides for joining lines together.
     /// </summary>
-    internal class JoinLinesCommand : BaseCommand
+    internal sealed class JoinLinesCommand : BaseCommand
     {
+        #region Singleton
+
+        public static JoinLinesCommand Instance { get; private set; }
+
+        public static void Initialize(CodeMaidPackage package)
+        {
+            Instance = new JoinLinesCommand(package);
+            package.SettingsMonitor.Watch(s => s.Feature_JoinLines, Instance.Switch);
+        }
+
+        #endregion Singleton
+
         #region Fields
 
         private readonly UndoTransactionHelper _undoTransactionHelper;
@@ -22,10 +34,9 @@ namespace SteveCadwallader.CodeMaid.Integration.Commands
         /// </summary>
         /// <param name="package">The hosting package.</param>
         internal JoinLinesCommand(CodeMaidPackage package)
-            : base(package,
-                   new CommandID(PackageGuids.GuidCodeMaidCommandJoinLines, PackageIds.CmdIDCodeMaidJoinLines))
+            : base(package, PackageGuids.GuidCodeMaidMenuSet, PackageIds.CmdIDCodeMaidJoinLines)
         {
-            _undoTransactionHelper = new UndoTransactionHelper(package, "CodeMaid Join");
+            _undoTransactionHelper = new UndoTransactionHelper(package, Resources.CodeMaidJoin);
         }
 
         #endregion Constructors