+++ CodeMaid/Model/Comments/CommentFormatter.cs
@@ -14,30 +14,30 @@ namespace SteveCadwallader.CodeMaid.Model.Comments
     {
         #region Fields
 
+        private readonly FormatterOptions _formatterOptions;
+        private readonly CommentOptions _commentOptions;
+
         private StringBuilder _builder;
-        private string _commentPrefix;
         private int _commentPrefixLength;
         private int _currentPosition;
         private bool _isAfterCommentPrefix;
         private bool _isFirstWord;
         private bool _isIndented;
-        private Regex _regex;
-        private int _tabSize;
 
         #endregion Fields
 
         #region Constructors
 
-        public CommentFormatter(ICommentLine line, string commentPrefix, int tabSize, Regex regex)
+        public CommentFormatter(ICommentLine line, FormatterOptions formatterOptions, CommentOptions commentOptions)
         {
+            _formatterOptions = formatterOptions;
+            _commentOptions = commentOptions;
+
             _builder = new StringBuilder();
             _currentPosition = 0;
-            _regex = regex;
-            _tabSize = tabSize;
             _isFirstWord = true;
             _isIndented = false;
-            _commentPrefix = commentPrefix ?? string.Empty;
-            _commentPrefixLength = WordLength(_commentPrefix);
+            _commentPrefixLength = WordLength(commentOptions.Prefix);
 
             // Special handling for the root XML line, it should not output it's surrounding xml
             // tags, only it's child lines.
@@ -130,7 +130,7 @@ namespace SteveCadwallader.CodeMaid.Model.Comments
                 _currentPosition = 0;
             }
 
-            _builder.Append(_commentPrefix);
+            _builder.Append(_commentOptions.Prefix);
             _currentPosition += _commentPrefixLength;
             _isFirstWord = true;
             _isIndented = false;
@@ -160,7 +160,7 @@ namespace SteveCadwallader.CodeMaid.Model.Comments
             }
             else if (line.Content != null)
             {
-                var matches = _regex.Matches(line.Content).OfType<Match>().Select(x => new CodeCommentMatch(x)).ToList();
+                var matches = _commentOptions.Regex.Matches(line.Content).OfType<Match>().Select(x => new CodeCommentMatch(x, _formatterOptions)).ToList();
 
                 // Remove empty matches from the start and end of the comment.
                 CodeCommentMatch m;
@@ -216,10 +216,6 @@ namespace SteveCadwallader.CodeMaid.Model.Comments
                 {
                     NewLine();
                 }
-                else if (!_isFirstWord && Settings.Default.Formatting_CommentXmlSpaceTags)
-                {
-                    Append(CodeCommentHelper.Spacer);
-                }
 
                 // Always consider the word after the opening tag as the first word to prevent an
                 // extra space before.
@@ -227,7 +223,7 @@ namespace SteveCadwallader.CodeMaid.Model.Comments
 
                 foreach (var match in matches)
                 {
-                    if (match.IsList)
+                    if (match.IsLiteral || match.IsList)
                     {
                         if (!_isFirstWord)
                         {
@@ -235,10 +231,14 @@ namespace SteveCadwallader.CodeMaid.Model.Comments
                         }
 
                         Indent(indentLevel);
+                    }
+
+                    if (match.IsList)
+                    {
                         Append(match.ListPrefix);
 
-                        // List items include their spacing and do not require additional space,
-                        // thus we are logically still on the first word.
+                        // List items include their spacing and do not require additional space, thus
+                        // we are logically still on the first word.
                         _isFirstWord = true;
                     }
 
@@ -274,8 +274,8 @@ namespace SteveCadwallader.CodeMaid.Model.Comments
                                 NewLine();
                                 Indent(indentLevel);
 
-                                // If linewrap is on a list item, add extra spacing to align the
-                                // text with the previous line.
+                                // If linewrap is on a list item, add extra spacing to align the text
+                                // with the previous line.
                                 if (match.IsList)
                                 {
                                     Append(string.Empty.PadLeft(WordLength(match.ListPrefix), CodeCommentHelper.Spacer));
@@ -303,11 +303,6 @@ namespace SteveCadwallader.CodeMaid.Model.Comments
 
                         NewLine(true);
                     }
-                }
-
-                if (!forceBreak && Settings.Default.Formatting_CommentXmlSpaceTags)
-                {
-                    Append(CodeCommentHelper.Spacer);
                 }
 
                 if (_currentPosition == 0 || _currentPosition > _commentPrefixLength && forceBreak)
@@ -355,7 +350,7 @@ namespace SteveCadwallader.CodeMaid.Model.Comments
             if (isCopyrightTag && Settings.Default.Formatting_CommentXmlValueIndent < 1)
             {
                 _commentPrefixLength += CodeCommentHelper.CopyrightExtraIndent;
-                _commentPrefix += string.Empty.PadLeft(CodeCommentHelper.CopyrightExtraIndent);
+                _commentOptions.Prefix += string.Empty.PadLeft(CodeCommentHelper.CopyrightExtraIndent);
             }
 
             // Increase the indent level.
@@ -365,6 +360,11 @@ namespace SteveCadwallader.CodeMaid.Model.Comments
 
             // If true the tag should be alone on it's own line.
             tagOnOwnLine |= isLiteralContent;
+
+            if (!tagOnOwnLine && Settings.Default.Formatting_CommentXmlSpaceTags)
+            {
+                Append(CodeCommentHelper.Spacer);
+            }
 
             // If the literal content of an XML tag is set, output that content without formatting.
             if (isLiteralContent)
@@ -392,8 +392,8 @@ namespace SteveCadwallader.CodeMaid.Model.Comments
                 {
                     if (needBreakBefore) NewLine();
 
-                    // Parse function returns true if it had to wrap lines. If so, we need to force
-                    // a newline before the closing tag.
+                    // Parse function returns true if it had to wrap lines. If so, we need to force a
+                    // newline before the closing tag.
                     needBreakBefore = Parse(l, indentLevel, xmlTagLength);
                     tagOnOwnLine |= needBreakBefore;
                 }
@@ -405,16 +405,20 @@ namespace SteveCadwallader.CodeMaid.Model.Comments
             if (isCopyrightTag && Settings.Default.Formatting_CommentXmlValueIndent < 1)
             {
                 _commentPrefixLength -= CodeCommentHelper.CopyrightExtraIndent;
-                _commentPrefix = _commentPrefix.Substring(0, _commentPrefixLength);
+                _commentOptions.Prefix = _commentOptions.Prefix.Substring(0, _commentPrefixLength);
             }
 
             // If opening tag was on own line, do the same for the closing tag.
             if (tagOnOwnLine && !_isFirstWord)
             {
                 NewLine();
+                Indent(indentLevel);
+            }
+            else if (Settings.Default.Formatting_CommentXmlSpaceTags)
+            {
+                Append(CodeCommentHelper.Spacer);
             }
 
-            Indent(indentLevel);
             Append(line.CloseTag);
 
             return tagOnOwnLine || CommentLineXml.SingleLineElementNames.Contains(line.TagName, StringComparer.OrdinalIgnoreCase);
@@ -458,7 +462,7 @@ namespace SteveCadwallader.CodeMaid.Model.Comments
         /// <returns>The length of the string.</returns>
         private int WordLength(string word)
         {
-            return word == null ? 0 : word.Length + word.Count(c => c == '\t') * (_tabSize - 1);
+            return word == null ? 0 : word.Length + word.Count(c => c == '\t') * (_formatterOptions.TabSize - 1);
         }
 
         #endregion Methods