+++ CodeMaid/Model/Comments/CommentMatch.cs
@@ -8,7 +8,7 @@ namespace SteveCadwallader.CodeMaid.Model.Comments
     {
         #region Constructors
 
-        public CodeCommentMatch(Match match)
+        public CodeCommentMatch(Match match, FormatterOptions formatterOptions)
         {
             if (!match.Success)
             {
@@ -16,12 +16,23 @@ namespace SteveCadwallader.CodeMaid.Model.Comments
                 return;
             }
 
-            Indent = match.Groups["indent"].Success ? match.Groups["indent"].Value.Length : 0;
-            ListPrefix = match.Groups["listprefix"].Success ? match.Groups["listprefix"].Value : null;
-            Words = match.Groups["words"].Success ? match.Groups["words"].Captures.OfType<Capture>().Select(c => c.Value).ToList() : null;
+            if (formatterOptions.IgnoreTokens.Any(p => match.Value.StartsWith(p)))
+            {
+                Words = new List<string> { match.Groups["line"].Value };
+                IsLiteral = true;
+                IsEmpty = false;
+                IsList = false;
+            }
+            else
+            {
+                Indent = match.Groups["indent"].Success ? match.Groups["indent"].Value.Length : 0;
+                ListPrefix = match.Groups["listprefix"].Success ? match.Groups["listprefix"].Value : null;
+                Words = match.Groups["words"].Success ? match.Groups["words"].Captures.OfType<Capture>().Select(c => c.Value).ToList() : null;
 
-            IsEmpty = string.IsNullOrWhiteSpace(match.Value) || Words == null || Words.Count < 1;
-            IsList = !string.IsNullOrWhiteSpace(ListPrefix);
+                IsLiteral = false;
+                IsEmpty = string.IsNullOrWhiteSpace(match.Value) || Words == null || Words.Count < 1;
+                IsList = !string.IsNullOrWhiteSpace(ListPrefix);
+            }
 
             // In the case of a list prefix but no content (e.g. hyphen line) convert to regular content.
             if (IsEmpty && IsList)
@@ -42,6 +53,8 @@ namespace SteveCadwallader.CodeMaid.Model.Comments
         public bool IsEmpty { get; private set; }
 
         public bool IsList { get; private set; }
+
+        public bool IsLiteral { get; private set; }
 
         public int Length
         {
@@ -79,6 +92,9 @@ namespace SteveCadwallader.CodeMaid.Model.Comments
                 return false;
 
             if (IsList && other.Indent < 1)
+                return false;
+
+            if (IsLiteral || other.IsLiteral)
                 return false;
 
             foreach (var word in other.Words)