+++ GlobalAssemblyInfo.cs
@@ -5,11 +5,11 @@ using System.Resources;
 using System.Runtime.InteropServices;
 
 [assembly: AssemblyCompany(Vsix.Author)]
-[assembly: AssemblyCopyright("Copyright 2007-2017 Steve Cadwallader (LGPL v3)")]
+[assembly: AssemblyCopyright("Copyright 2007-2018 Steve Cadwallader (LGPL v3)")]
 [assembly: AssemblyDescription(Vsix.Description)]
 [assembly: AssemblyFileVersion(Vsix.Version)]
 [assembly: AssemblyProduct(Vsix.Name)]
 [assembly: AssemblyVersion(Vsix.Version)]
 [assembly: CLSCompliant(false)]
 [assembly: ComVisible(false)]
-[assembly: NeutralResourcesLanguage("en-US")]
+[assembly: NeutralResourcesLanguage("en-US", UltimateResourceFallbackLocation.Satellite)]